+++ packages/sdk/src/contract-functions/signPermitChallenge/index.ts
@@ -1,0 +1,5 @@
+export {
+  type ISignPermitChallengeParams,
+  type ISignPermitChallengeResult,
+  signPermitChallenge,
+} from './signPermitChallenge';